Only build bench for the master branch (#2584)
The previous PR enables the bench_dbms
https://github.com/PingCAP-QE/ci/pull/2515 to ensure the new code
changes do not break building the bench binary.
But in the old branches, the bench binary have been broken for a long
days and it brings trouble for cherry-pick to LTS branches and hotfix
branches.

This PR only build the bench_dbms on the master branch. Ignore it for
the LTS branches and hotfix branches.