feat: build tidb server without race (#2602)
Build tidb-server which not enable race in mysql-test pipelines.

Comparison of test time before and after starting the race.
* race  about 30 minutes
* without abount 12 minutes

Now, due to the presence of cache binary logic in pipelines, even though
the build command is enabled race `WITH_RACE=1 make server`, it is
actually using a tidb-server without race enabled for testing.