chore: adjust timeout for tidb ci (#2683)
Currently, Bazel will retry three times by default. If the timeout is
set to "long", each timeout limit for each attempt is 15 minutes. If
there are timeout issues with the case, when it reaches three timeouts,
the entire task will be terminated will cause the test logs to not be
output to stdout.

So here we will increase the timeout for the entire task by five
minutes.

PS: At the same time, it would be better to adjust the bazel timeout
(`long` -> `moderate `) for testing.