feat: support checkout tidb-test with pr in batch merge (#2838)
When checking out the tidb-test code, it is necessary to determine
whether the tidb-test PR has been specified based on the tidb pr title.
During batch merge, multiple tidb PRs will be processed simultaneously.
At this checkout step, we need to obtain all tidb pr titles to determine
if the tidb-test PR has been specified.

Which situation is not allowed?
during batch merge, it is not allowed to specify the tidb-test branch or
commit in the pr title for checking out code.