feat: lightning support run_if_changed (#2891)
The lightning pipeline supports triggering based on conditions. If code
changes are detected in the corresponding directory, the corresponding
test pipeline will be triggered. Block PR merging if the test pipeline
execution fails.

It should be noted that because prow does not support configuring these
two parameters at the same time (run_if_changed and
skip_if_only_changed), after applying this change. If there are changes
in pure document categories for lightning, it will also trigger the test
pipeline.