chore(jenkins): remove the deprecated CI jobs (#3235)
This pull request removes the Jenkins pipeline jobs and associated
scripts for the `importer` project. The changes include the deletion of
job definitions and build/test scripts.

### Removal of Jenkins pipeline jobs:
*
[`jenkins/jobs/ci2/tikv/importer/ghpr_build.groovy`](diffhunk://#diff-0af8fec9836b56ac2cd2bfb01076ebde0df7147bbf8f52f2ae45449de596f810L1-L79):
Removed the job definition for `importer_ghpr_build` that included
parameters, properties, and triggers.
*
[`jenkins/jobs/ci2/tikv/importer/ghpr_test.groovy`](diffhunk://#diff-ca788a484622a307b4ae4ab5551f2cf619508fb7a944943efdbc95d1507b3131L1-L79):
Removed the job definition for `importer_ghpr_test` that included
parameters, properties, and triggers.

### Removal of build and test scripts:
*
[`jenkins/pipelines/ci/importer/importer_ghpr_build.groovy`](diffhunk://#diff-bf3b3367e924f18f19d1992195826ea9346530c73c8f4555f2943ba784cc39bbL1-L70):
Deleted the build script for `importer_ghpr_build`, which included
stages for preparation and testing, as well as Slack notifications.
*
[`jenkins/pipelines/ci/importer/importer_ghpr_test.groovy`](diffhunk://#diff-4186fc216d52caed75fb38460f09f9e3036b690e4789eeb5c17b33a04beec5ebL1-L80):
Deleted the test script for `importer_ghpr_test`, which included stages
for preparation and testing, as well as Slack notifications.