chore(tiflow): keep the trigger command and context consistent (#3308)
To reduce the memory cost of triggering commands, keep the trigger
command consistent with the context.

Taking pull-cdc-integration-kafka-test as an example. Context of this
pipeline is `pull-cdc-integration-kafka-test`
* The current trigger command is `/test cdc-integration-kafka-test`
* The adjusted trigger command is `/test
pull-cdc-integration-kafka-test`

For `jenkins-ticdc/verify`, because this pipeline includes unit tests
and build for multiple components (CDC and other components), so we
update this context to `pull-verify`
* The context will change from `jenkins-ticdc/verify` to `pull-verify`
* The trigger command will change from `/test verify` to `/test
pull-verify`

---------

Signed-off-by: purelind <purelind@gmail.com>