fix(cd): update docker image repo and dockerfiles for devbuild job (#3335)
This pull request includes significant changes to the
`jenkins/pipelines/cd/dev-build.groovy` file to improve the mapping and
retrieval of Docker image URLs and Dockerfile URLs. The most important
changes include the introduction of a new mapping for Docker image
repositories, updates to the Dockerfile URL retrieval logic, and
adjustments to the image naming conventions.

Improvements to Docker image and Dockerfile mappings:

*
[`jenkins/pipelines/cd/dev-build.groovy`](diffhunk://#diff-883dea4e305e8d8dffff92a73f812d069fd76eaf118e96b9efd1bdc219de9fe6L1-R35):
Introduced `DockerImgRepoMapping` to map products to their respective
Docker image repositories. This replaces the previous `DockerMapping`
and expands the coverage to include more products.
*
[`jenkins/pipelines/cd/dev-build.groovy`](diffhunk://#diff-883dea4e305e8d8dffff92a73f812d069fd76eaf118e96b9efd1bdc219de9fe6L26-R101):
Updated the `get_dockerfile_url` function to use a more comprehensive
mapping for retrieving Dockerfile URLs based on the product and version.
The new mapping covers both versions >= 'v6.6.0' and < 'v6.6.0'.

Adjustments to image naming conventions:

*
[`jenkins/pipelines/cd/dev-build.groovy`](diffhunk://#diff-883dea4e305e8d8dffff92a73f812d069fd76eaf118e96b9efd1bdc219de9fe6L90-R152):
Changed the `ProductForDocker` assignment to use `DockerImgRepoMapping`
instead of the old `DockerMapping`. This ensures consistency with the
new Docker image repository mapping.
*
[`jenkins/pipelines/cd/dev-build.groovy`](diffhunk://#diff-883dea4e305e8d8dffff92a73f812d069fd76eaf118e96b9efd1bdc219de9fe6L99-R161):
Modified the image naming logic for hotfixes to remove the `qa` prefix
from the image repository path.

---------

Signed-off-by: wuhuizuo <wuhuizuo@126.com>