feat(tiflow): download exact version content in hotfix branch (#3380)
Over time, the release branch will be continuously updated, and for the
hotfix branch (from the tag), using non-fixed artifacts in integration
testing may lead to test failures due to incompatible changes in
dependency components(tidb & tikv & pd & tiflash). Therefore, we handle
hotfixes separately in the pipeline; when a hotfix branch is detected,
we download and test the dependencies for that tag version.

---------

Signed-off-by: purelind <purelind@gmail.com>