feat(tiflow): download exact version content for tiflow hotfix (#3394)
Download exact version content for tiflow hotfix branch integration
tests.
Over time, the release branch will be continuously updated, and for the
hotfix branch (from the tag), using non-fixed artifacts in integration
testing may lead to test failures due to incompatible changes in
dependency components(tidb & tikv & pd & tiflash). Therefore, we handle
hotfixes separately in the pipeline; when a hotfix branch is detected,
we download and test the dependencies for that tag version.
1. current PR will support for those lts hotfix branch: 6.1 & 6.5 & 7. 1
& 7.5 & 8.5
2. Support for 8.1 is in this PR:
https://github.com/PingCAP-QE/ci/pull/3380