chore(scripts/flow/ga): do not create github version tag on tidb-tools repo for since v9.0.0 (#3402)
This pull request includes an update to the
`scripts/flow/ga/create_github_release.ts` file to exclude the
`tidb-tools` repository from releases starting from version 9.0.0. This
change ensures that the `tidb-tools` repository, which has migrated its
code to the `tiflow` repository, is no longer included in the release
process.

*
[`scripts/flow/ga/create_github_release.ts`](diffhunk://#diff-c5760f7bc08ebe7f0db7f1eee0626e0844179cd9e4def16db9c809f45d5e4e89R111-R116):
Added logic to filter out the `tidb-tools` repository for versions 9.0.0
and above.

Signed-off-by: wuhuizuo <wuhuizuo@126.com>