feat(ticdc): modify golang and net-tool cpu for kafka light pipelines (#3453)
In the monitoring of Storage Light Pipelines, the CPU throttling metric
for net-tool is consistently around 56%. Therefore, it is necessary to
increase the CPU allocation for net-tool. The CPU usage of Golang
doesn’t reach 6 cores, so the CPU resources can be reduced a bit. This
PR increases net-tool’s CPU from 100m to 400m, and decreases golang's
CPU from 6 to 4.