chore(pingcap/tidb-operator): update dind-daemon configuration in latest-presubmits.yaml
This commit modifies the `latest-presubmits.yaml` file to change the dind-daemon container's image to `docker:28.1-dind`, updates the command to `dockerd-entrypoint.sh`, and adjusts the environment variable settings. Additionally, it changes the mount path for the Docker socket from `/run/user/1001` to `/var/run`, streamlining the configuration for better compatibility.