chore(pingcap/tidb): rename pull_vector_search_test to  pull_tiflash_integration_test (#3545)
This commit introduces a new Jenkins pipeline job for the Tiflash
integration test, defined in `pull_tiflash_integration_test.groovy`. It
includes a pod specification in `pod-pull_tiflash_integration_test.yaml`
with resource requests and limits for the Golang and network tool
containers. Additionally, the `latest-presubmits.yaml` file is updated
to register the new job, ensuring it can be triggered and rerun as
needed.