fix(pipelines/tiflow): update binary download script references to specific release versions on hotfix branch (#3616)
Updated the binary download script references in multiple integration
test files to use the specific release version (release-7.1,
release-8.1, release-8.5) instead of the base reference. This change
ensures that the correct binaries are downloaded for the respective
release branches.

This is to ensure that there are no errors when executing downloads for
the hotfix branch; in fact, the hotfix branch will download dependencies
with precise tags.