chore(pingcap/tidb): remove Go version check from build pipelines (#3686)
Revert #3683

The original intention was to ensure that the tidb and plugin GitHub
repositories use the same Go version in their go.mod files.

This inconsistency issue manifests itself in our CI tasks on the day or
the next day of each Go patch release. This isn't an environmental
issue, but an engineering issue of TiDB repository!

Symptoms similar to this:
```log
[2025/08/07 06:51:03.979 +00:00] [FATAL] [terror.go:309] ["unexpected error"] [error="plugin.Open(\"plugin-so/audit-1\"): plugin was built with a different version of package runtime"]
```


**Currently, some contributors do not understand this requirement and
feel that this step undermines their work. If you agree, you can add
your approval. After this rollback, this consistency guarantee must be
manually checked by contributors or reviewers.**