feat(pipelines/tidb): update pull-error-log-review job configuration of tidb (#3691)
- Changed the `always_run` property to `true` for the
`pull-error-log-review` job, ensuring it runs on every trigger.
- Added `skip_report: true` to prevent reporting on this job,
streamlining the CI process.

Enable automatic triggering. To verify the effectiveness of the matching
rules, please note that this task currently has status reporting turned
off, so task cost failures will not be reflected on the PR GitHub page
and will not affect PR merging.