fix(prow-jobs/tikv/pd): fix dependency commit check (#3735)
This pull request makes minor improvements to the dependency checking
script in the `prow-jobs/tikv/pd/common-presubmits.yaml` file. The
changes enhance logging and improve the accuracy of commit verification
for Go module dependencies.

Dependency checking improvements:

* Added a log message to output the commit hash of each Go module
dependency when using a pseudo-version, making it easier to trace which
commit is being checked.
* Improved the logic for verifying if a commit is in the expected branch
by comparing only the first 12 characters of the commit hash, which
matches the format used by Go pseudo-versions.