feat(scripts/artifacts): add support for downloading pd-ctl and fix downloading tikv-worker (#3843)
## Feat: Add support for downloading pd-ctl artifact

This pull request introduces the capability to download the `pd-ctl`
binary from OCI registries. This is a necessary addition to ensure that
all required components for TiDB cluster management can be provisioned
consistently.

The following changes have been made:

* Added a new command-line argument `-pd-ctl` (and its long form
`--pd-ctl`) to specify the version of `pd-ctl` to download.
* Implemented logic within the `main` function to download and extract
the `pd-ctl` artifact if the `PD_CTL` environment variable is set.
* Constructed the OCI URL for `pd-ctl` using the provided version and
the configured registry host.
* Updated the `parse_cli_args` function to handle the new `-pd-ctl`
argument.
* Included `PD_CTL` in the output of the script's argument parsing for
better visibility.

This enhancement allows users to explicitly specify and download
`pd-ctl` alongside other TiDB components, streamlining the setup and
deployment process.

## Fix: Fix downloading tikv-worker artifact

This pull request fix the downloading tag of tikv-worker.

* Use the value of tikv-worker argument as the tag for downloading.

close #3842

---------

Signed-off-by: tenfyzhong <tenfy@tenfy.cn>