Update pipeline to use separate pod templates for build and test stages (#3844)
This pull request updates the pod template file paths for the
`pingcap/ticdc/latest/pull_cdc_kafka_integration_light_next_gen`
pipeline.

The changes are as follows:

*   Updated `POD_TEMPLATE_FILE` to point to `pod-test.yaml`.
*   Updated `POD_TEMPLATE_FILE_BUILD` to point to `pod-build.yaml`.

This change is necessary to differentiate between pod configurations
used for testing and those used for building, allowing for more specific
and controlled resource allocation for each stage of the pipeline.

Signed-off-by: tenfyzhong <tenfy@tenfy.cn>