Add release-nextgen branch to presubmit jobs to ensure compatibility with upcoming releases. (#3888)
This PR updates the branch configuration for two presubmit jobs in the
TiCDC project. The goal is to ensure these jobs run on newly created
`release-nextgen-*` branches, in addition to the existing `master`
branch. This will allow for better testing and validation of features
being developed for the next generation of TiCDC as they are merged into
these release branches.

Changes include:

* Added `^release-nextgen-\d+$` to the `branches` field for the `check`
presubmit job.
* Added `^release-nextgen-\d+$` to the `branches` field for the `build`
presubmit job.

No issues are closed by this change.

Signed-off-by: tenfyzhong <tenfy@tenfy.cn>