fix(tekton/v1): fix image sync cronjob (#4023)
This pull request makes a minor update to the Tekton pipeline
configuration for the `env-prod2` environment. The main change is an
update to the image list in the `fetch-sync-pre-ver-image-tags-to-hub`
cronjob, and a small cleanup in the kustomization file.

- **Image List Update:**
- Updated the `ocis` array in
`fetch-sync-pre-ver-image-tags-to-hub.yaml` to use `pingcap/ticdc/image`
instead of the deprecated `pingcap/images/cdc` image.

- **Configuration Cleanup:**
- Removed a duplicate or unnecessary reference to
`harbor/cronjobs/fetch-sync-pre-ver-image-tags-to-hub.yaml` in the
`kustomization.yaml` resources list.